fallback when tau cannot be estimated
When dataset is too small or ill-conditioned, the average point distance might fail to be estimated. We fall back to largest average distance we found in the sample. There is likely caveat in this choice and reverting back to a metric tree is probably a better choice. Added optimization that tau is never estimated if rho is zero (i.e. only metric trees are built)